.envのデフォルトエンジン情報を読み出す部分を共通化する
