+++ src/backend/browser/contract.ts
@@ -1,8 +1,7 @@
-import { type EngineInfo, envEngineInfoSchema } from "@/type/preload";
+import { loadEnvEngineInfos } from "@/domain/defaultEngine/envEngineInfo";
+import { type EngineInfo } from "@/type/preload";
 
-const baseEngineInfo = envEngineInfoSchema
-  .array()
-  .parse(JSON.parse(import.meta.env.VITE_DEFAULT_ENGINE_INFOS))[0];
+const baseEngineInfo = loadEnvEngineInfos()[0];
 
 export const defaultEngine: EngineInfo = (() => {
   const { protocol, hostname, port, pathname } = new URL(baseEngineInfo.host);
