+++ src/domain/defaultEngine/latetDefaultEngine.ts
@@ -1,5 +1,5 @@
 /**
- * デフォルトエンジン関連のモジュール
+ * デフォルトエンジンの最新情報関連のモジュール
  */
 
 import { z } from "zod";
