+++ tests/unit/domain/defaultEngine/defaultEngine.node.spec.ts
@@ -2,7 +2,7 @@
 
 import path from "path";
 import fs from "fs";
-import { fetchDefaultEngineInfos } from "@/domain/defaultEngine";
+import { fetchDefaultEngineInfos } from "@/domain/defaultEngine/defaultEngine";
 
 const currentDir = "tests/unit/domain/defaultEngine";
